Contour/requirements.txt
torch==2.4.1
torchvision==0.19.1
torchaudio==2.4.1
--extra-index-url https://download.pytorch.org/whl/cu121

transformers==4.55.0
accelerate>=0.33.0
datasets>=2.20.0
safetensors>=0.4.3
pillow==10.4.0
matplotlib==3.8.4
tqdm==4.66.4
numpy>=1.26.0
scipy>=1.11.4
psutil>=5.9.0
opencv-python-headless>=4.9.0.80

